requirements.txt
# Google Drive API dependencies
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# AutoGen framework for AI agents
autogen-agentchat==0.5.7
autogen-core==0.5.7
autogen-ext==0.5.7

# Image processing
Pillow==10.4.0

# Additional dependencies for AutoGen
openai==1.54.4
pydantic==2.10.3
typing-extensions==4.12.2 